Uncompleted work

+++ prototype/lib/net/allocatorutil.rb
@@ -15,7 +15,7 @@ def parse_options(argv)
   parser.on("-c", "--config=FILE", "job configuration file", String) { |config| options[:config] = config }
   parser.on("-u", "--user=USER", "user ID, which used in job package routing", String) { |user| options[:user_id] = user }
   parser.on("-t", "--timeout=SEC", "timeout in seconds, between consecutive connection tries",
-            "default=#{options[:timeout]}", Integer) { |timeout| options[:timeout] = timeout }
+            "default=#{options[:timeout]}", Float) { |timeout| options[:timeout] = timeout }
   parser.on("-r", "--reruns=TIMES", "how many times try to connect to server",
             "default=#{options[:reruns]}", Integer) { |reruns| options[:reruns] = reruns }
